requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
werkzeug==2.3.6
python-dotenv==1.0.0
email_validator==2.0.0.post2